% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{exclude_accel}
\alias{exclude_accel}
\title{Remove days with too few/too much weartime and NHANES data quality flags.}
\usage{
exclude_accel(act, flags, threshold = 600, rm_PAXSTAT = TRUE,
  rm_PAXCAL = TRUE, return_act = FALSE)
}
\description{
This function subsets accelerometry data by wear/non-wear time criteria, as well as
NHANES data quality flags.
}
\examples{
\dontrun{
process_accel(write=FALSE)
}

}
\references{

}
