% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesdata-package.R
\docType{package}
\name{nhanesdata-package}
\alias{nhanesdata}
\alias{nhanesdata-package}
\title{nhanesdata: Harmonized Access to NHANES Survey Data}
\description{
Instant access to harmonized National Health and Nutrition Examination Survey (NHANES) data spanning 1999-2023. Retrieve pre-processed datasets from reliable cloud storage with automatic type reconciliation and integrated search tools for variables and datasets. Simplifies NHANES data workflows by handling cycle management and maintaining data consistency across survey waves. Data is sourced from \url{https://www.cdc.gov/nchs/nhanes/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kyleGrealis/nhanesdata}
  \item \url{https://www.kylegrealis.com/nhanesdata/}
  \item Report bugs at \url{https://github.com/kyleGrealis/nhanesdata/issues}
}

}
\author{
\strong{Maintainer}: Kyle Grealis \email{kylegrealis@proton.me} (\href{https://orcid.org/0000-0002-9223-8854}{ORCID})

Other contributors:
\itemize{
  \item Amrit Baral \email{abaral3@jhu.edu} (\href{https://orcid.org/0000-0002-0961-9665}{ORCID}) [contributor]
  \item Natalie Neugaard \email{nataliegoulett@gmail.com} (\href{https://orcid.org/0009-0008-0820-5111}{ORCID}) [contributor]
  \item Raymond Balise \email{balise@miami.edu} (\href{https://orcid.org/0000-0002-9856-5901}{ORCID}) [contributor]
  \item Johannes Thrul \email{jthrul@jhu.edu} (\href{https://orcid.org/0000-0001-8929-9579}{ORCID}) [contributor]
  \item Janardan Devkota \email{jdevkot1@jhu.edu} (\href{https://orcid.org/0009-0007-6734-777X}{ORCID}) [contributor]
}

}
\keyword{internal}
