% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesdata.R
\docType{package}
\name{nhanesdata}
\alias{nhanesdata}
\alias{nhanesdata-package}
\title{Data from Continuous NHANES (Years 1999-2016)}
\description{
Contains data from the National Health and Nutrition Examination Survey,
years 1999-2016. Current as of Jan. 8, 2019.
}
\details{
\tabular{ll}{
Package: \tab nhanesdsata \cr
Type: \tab Package \cr
Version: \tab 1.1.1 \cr
Date: \tab 2019-02-08 \cr
License: \tab GPL-3 \cr
}

This package simply contains data from NHANES 1999-2016. Each dataset can be
loaded into R by running \code{data(filename)}, where filename is the file
name listed on the NHANES website
(\url{http://www.cdc.gov/nchs/nhanes/nhanes_questionnaires.htm}).
}
\examples{
# Load demographics from 1999-2000 and from 2001-2002
data(demo)
data(demo_b)


}
\references{
Centers for Disease Control and Prevention (CDC). National Center for Health
Statistics (NCHS). National Health and Nutrition Examination Survey Data.
Hyattsville, MD: US Department of Health and Human Services, Centers for
Disease Control and Prevention, 1999-2016.
\url{https://wwwn.cdc.gov/nchs/nhanes/Default.aspx}. Accessed Jan. 8, 2019.

Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen \cr \email{vandomed@gmail.com}
}
