% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{process_accel}
\alias{process_accel}
\title{Process NHANES 2003-2004 and 2005-2006 accelerometry data}
\usage{
process_accel(write = FALSE, local = FALSE, localpath = NULL,
  deleteraw = TRUE)
}
\arguments{
\item{write}{logical argument indicating whether a .rda file should be created for each wave of processed data.
Defaults to FALSE.}

\item{local}{logical argument indicating whether the zippped raw .xpt accelerometry files are stored locally.
If FALSE, will download the data into a temporary file from the CDC website and process the data. If TRUE,
the zipped data will be sourced locally. Defaults to FALSE.}

\item{localpath}{character string indicating where the locally zipped raw .xpt files are.}

\item{deleteraw}{logical argument indicating whether to delete the unzipped .xpt files after reading them into R.}
}
\description{
This function processes the raw NHANES 2003-2004 and 2005-2006 accelerometry data
as provided by the NHANES study on the CDC website.
Note that due to the large file size of the unzipped .xpt files, this function uses a non-trivial
amount of RAM (~12 GB peak). To avoid crashing your computer when running, please ensure you have
enough RAM available.
}
\examples{
\dontrun{
library("nhanesaccel")
process_accel()
process_flags()
}

}
