% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{process_covar}
\alias{process_covar}
\title{Process covariate data for NHANES 2003-2004 and 2005-2006 waves}
\usage{
process_covar(write = FALSE)
}
\arguments{
\item{write}{logical argument indicating whether a csv file of wear/non-wear flags
should be created for each wave of processed data. Defaults to FALSE.}
}
\description{
This function merges covariate data
from several NHANES data files.
The resulting data matrix includes survey weights as well as a number of demographic and
comorbidity responses.
}
\examples{
\dontrun{


}

}
\references{

}
