% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{process_flags}
\alias{process_flags}
\title{Process wear/non-wear flags for NHANES 2003-2004 and 2005-2006 accelerometry data}
\usage{
process_flags(write = FALSE, local = FALSE, localpath = NULL,
  window = 90, tol = 2, tol.upper = 99, ...)
}
\arguments{
\item{write}{logical argument indicating whether a .rda file of wear/non-wear flags
should be created for each wave of processed data. Defaults to FALSE.}

\item{local}{logical argument indicating whether the processed .rda accelerometry files are stored locally.
If FALSE, will load them from the package data.}

\item{localpath}{character string indicating where the locally processed .rda files are if you don't want to use those processed and included
in the package.}

\item{window}{size of the moving window used to assess non-wear in minutes. Defaults to 90 minutes.
See \code{\link{accel.weartime}} for more details.}

\item{tol}{maximum number of minutes with counts greater than 0 within the window allowed before a particular
minute is considered "wear". That is, if for a given minute, the window around that minute has tol + 1
activity counts greater than 0, this miute is considered "wear".
See \code{\link{accel.weartime}} for more details.}

\item{tol.upper}{maximium activity count for any minute within the window
Defaults to 99. That is, for a given minute, if the window contains any minutes with
activity counts greater than tol.upper, this minute is considered "wear".
See \code{\link{accel.weartime}} for more details.}

\item{...}{aditional arguments to be passed to \code{\link{accel.weartime}}.}
}
\description{
This function creates wear/non-wear flag matrices for processed NHANES 2003-2004 and 2005-2006 accelerometry data
}
\examples{
\dontrun{
process_accel(write=FALSE)
}

}
\references{

}
