% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{reweight_accel}
\alias{reweight_accel}
\title{Reweight NHANES accelerometry data}
\usage{
reweight_accel(data)
}
\description{
This function re-weights accelerometry data for NHANES 2003-2004,2005-2006 waves
}
\examples{
\dontrun{
process_accel(write=FALSE)
}

}
\references{

}
