% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{process_accel}
\alias{process_accel}
\title{Process NHANES 2003-2004 and 2005-2006 accelerometry data}
\usage{
process_accel(write = FALSE)
}
\arguments{
\item{write}{logical argument indicating whether a csv file should be created for each wave of processed data.
Defaults to FALSE.}
}
\description{
This function processes the raw NHANES 2003-2004 and 2005-2006 accelerometry data
as provided by the NHANES study on the CDC website.
Note that due to the large file size of the unzipped .xpt files, this function uses a non-trivial
amount of RAM (~12 GB peak). To avoid crashing your computer when running, please ensure you have
enough RAM available.
}
\examples{
\dontrun{

process_accel(write=FALSE)

}

}
